requirements.txt
rembg
Pillow
Pyinstaller